+++ src/shared/ui/BaseComponent.ts
